% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmean_fn.R
\name{bca_wmean}
\alias{bca_wmean}
\title{following slides here: http://users.stat.umn.edu/~helwig/notes/bootci-Notes.pdf}
\usage{
bca_wmean(treat, covar, out, nboot, treat_form, out_levels, out_form,
  out_weights, out_model, wmean_est, alpha = 0.05)
}
\description{
following slides here: http://users.stat.umn.edu/~helwig/notes/bootci-Notes.pdf
}
