% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_fn.R
\name{compute_trt_spec_marg_cdf_ptwise_ci}
\alias{compute_trt_spec_marg_cdf_ptwise_ci}
\title{this function could be renamed as it's used for both CDF and PMF}
\usage{
compute_trt_spec_marg_cdf_ptwise_ci(pt_est, cov_est, alpha, cdf = TRUE)
}
\description{
this function could be renamed as it's used for both CDF and PMF
}
