% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_fn.R
\name{compute_trt_spec_marg_cdf_simul_ci}
\alias{compute_trt_spec_marg_cdf_simul_ci}
\title{this function could be renamed since it's used for CDF and PMF}
\usage{
compute_trt_spec_marg_cdf_simul_ci(pt_est, trt_spec_marg_cdf_eif,
  remove_last = TRUE, alpha)
}
\description{
this function could be renamed since it's used for CDF and PMF
}
