% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_fn.R
\name{compute_trt_spec_marg_dist_ptwise_ci}
\alias{compute_trt_spec_marg_dist_ptwise_ci}
\title{Compute simultaneous confidence interval for treatment-specific marginal distribution}
\usage{
compute_trt_spec_marg_dist_ptwise_ci(pt_est, cov_est, alpha, cdf = TRUE)
}
\arguments{
\item{pt_est}{The point estimate of the treatment-specific marginal CDF/PMF}

\item{cov_est}{Covariance matrix estimates.}

\item{alpha}{Confidence intervals have nominal level 1-\code{alpha}.}

\item{cdf}{Is this for CDF or PMF?}
}
\value{
Confidence interval
}
\description{
Compute simultaneous confidence interval for treatment-specific marginal distribution
}
