% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_fn.R
\name{compute_trt_spec_marg_dist_simul_ci}
\alias{compute_trt_spec_marg_dist_simul_ci}
\title{Compute simultaneous confidence interval for treatment-specific marginal distribution}
\usage{
compute_trt_spec_marg_dist_simul_ci(
  pt_est,
  trt_spec_marg_dist_eif,
  remove_last = TRUE,
  alpha
)
}
\arguments{
\item{pt_est}{The point estimate of the treatment-specific marginal CDF/PMF}

\item{trt_spec_marg_dist_eif}{The EIF estimates for the treatment-specific marginal 
CDF/PMF estimates}

\item{remove_last}{Should the last level be removed? Should be set equal to 
\code{TRUE} for CDF computations and \code{FALSE} for PMF computations.}

\item{alpha}{Confidence intervals have nominal level 1-\code{alpha}.}
}
\value{
Confidence interval
}
\description{
Compute simultaneous confidence interval for treatment-specific marginal distribution
}
