% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_estimation.R
\name{drord}
\alias{drord}
\title{Obtain estimates of ...}
\usage{
drord(out, treat, covar, out_levels = order(unique(out)),
  out_form = NULL, out_weights = rep(1, length(out_levels)),
  treat_form = "1", param = c("weighted-mean", "log-odds",
  "mann-whitney"), ci = c("bca", "wald"), alpha = 0.05,
  test = c("perm", "wald"), nboot = 10000, nperm = 10000,
  est_dist = FALSE, ...)
}
\description{
Obtain estimates of ...
}
\examples{
n <- 200
covar <- data.frame(x1 = runif(n), x2 = runif(n))
treat <- rbinom(n, 1, 1/2)
out <- rbinom(n, 3, plogis(-1 + covar$x1 - covar$x2 + 0.2 * treat))
out_levels <- 0:3
out_form <- "x1 + x2"
out_weights <- rep(1, 4)
treat_form <- "x1 + x2"
ci <- c("bca", "wald")
test <- c("perm", "wald")
alpha <- 0.05 
nboot <- 1e2
out_levels <- 0:3
}
