% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logodds_fn.R
\name{eif_theta_k}
\alias{eif_theta_k}
\title{get influence function for \theta_k}
\usage{
eif_theta_k(k, out, treat, trt_level, trt_spec_prob_est,
  trt_k_spec_cdf_est)
}
\description{
get influence function for \theta_k
}
