% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_fn.R
\name{estimate_cdf}
\alias{estimate_cdf}
\title{Map an estimate of the conditional PMF into an estimate of the conditional CDF}
\usage{
estimate_cdf(pmf_est)
}
\arguments{
\item{pmf_est}{A list of the treatment-specific PMF estimates}
}
\value{
A list of treatment-specific CDF estimates
}
\description{
Map an estimate of the conditional PMF into an estimate of the conditional CDF
}
