% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmean_fn.R
\name{estimate_cond_mean}
\alias{estimate_cond_mean}
\title{Map an estimate of treatment-specific PMF into an estimate of 
treatment specific conditional mean for each observation.}
\usage{
estimate_cond_mean(trt_spec_pmf_est, ordered_out_levels, ordered_out_weights)
}
\arguments{
\item{trt_spec_pmf_est}{The treatment-specific PMF estimates}

\item{ordered_out_levels}{Self explanatory}

\item{ordered_out_weights}{Self explanatory}
}
\value{
Vector of estimated conditional means
}
\description{
Map an estimate of treatment-specific PMF into an estimate of 
treatment specific conditional mean for each observation.
}
