% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mannwhitney_fn.R
\name{estimate_mannwhitney}
\alias{estimate_mannwhitney}
\title{Compute the estimate of Mann-Whitney based on conditional CDF and PMF}
\usage{
estimate_mannwhitney(cdf_est, pmf_est)
}
\arguments{
\item{cdf_est}{Conditional CDF estimates}

\item{pmf_est}{Conditional PMF estimates}
}
\value{
Mann-Whitney point estimate
}
\description{
Compute the estimate of Mann-Whitney based on conditional CDF and PMF
}
