% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_fn.R
\name{estimate_pmf}
\alias{estimate_pmf}
\title{Get a treatment-specific estimate of the conditional PMF. 
Essentially this is a wrapper function for \code{fit_trt_spec_reg}.}
\usage{
estimate_pmf(out, treat, covar, out_levels, out_form = NULL, out_model,
  treat_prob_est, stratify = FALSE, return_models = TRUE, ...)
}
\description{
Get a treatment-specific estimate of the conditional PMF. 
Essentially this is a wrapper function for \code{fit_trt_spec_reg}.
}
