% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_fn.R
\name{estimate_treat_prob}
\alias{estimate_treat_prob}
\title{Estimate probability of receiving each level of treatment
Assumes that \code{treat} comes in as a vector of 0's and 1's}
\usage{
estimate_treat_prob(treat, covar, treat_form)
}
\value{
A list where the first element is estimate of Pr(\code{treat} = 1 | \code{covar})
for \code{covar} equal to inputted values of \code{covar}
and second element is estimate of Pr(\code{treat} = 0 | \code{covar})
for \code{covar} equal to inputted values of \code{covar}
}
\description{
Estimate probability of receiving each level of treatment
Assumes that \code{treat} comes in as a vector of 0's and 1's
}
