% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logodds_fn.R
\name{evaluate_beta_cov}
\alias{evaluate_beta_cov}
\title{Get the covariance matrix for beta}
\usage{
evaluate_beta_cov(cdf_est, theta_cov)
}
\arguments{
\item{cdf_est}{Estimated CDFs}

\item{theta_cov}{Covariance matrix for CDF estimates}
}
\value{
Estimated covariance matrix for log-odds ratio parameters
}
\description{
Get the covariance matrix for beta
}
