% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mannwhitney_fn.R
\name{evaluate_mannwhitney_gradient}
\alias{evaluate_mannwhitney_gradient}
\title{Compute the estimated gradient of the Mann-Whitney parameter. Needed to derive 
standard error for Wald confidence intervals.}
\usage{
evaluate_mannwhitney_gradient(cdf_est, pmf_est)
}
\arguments{
\item{cdf_est}{Conditional CDF estimates}

\item{pmf_est}{Conditional PMF estimates}
}
\value{
3-length vector for delta method calculus
}
\description{
Compute the estimated gradient of the Mann-Whitney parameter. Needed to derive 
standard error for Wald confidence intervals.
}
