% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_fn.R
\name{evaluate_marg_cdf_eif}
\alias{evaluate_marg_cdf_eif}
\title{Get eif estimates for treatment-specific CDF}
\usage{
evaluate_marg_cdf_eif(cdf_est, treat_prob_est, treat, out, out_levels)
}
\arguments{
\item{cdf_est}{Estimated conditional CDF for \code{trt_level}.}

\item{treat}{A \code{numeric} vector containing treatment status. Should only assume 
a value 0 or 1.}

\item{out}{A \code{numeric} vector containing the outcomes.}

\item{out_levels}{A \code{numeric} vector containing all ordered levels of the 
outcome.}

\item{trt_prob_est}{Estimated propensity for \code{trt_level}.}
}
\value{
a list of eif estimates
}
\description{
Get eif estimates for treatment-specific CDF
}
