% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_fn.R
\name{evaluate_marg_cdf_simul_ci}
\alias{evaluate_marg_cdf_simul_ci}
\title{this function could be renamed since it's used for CDF and PMF
for CDF, set remove_last = FALSE, for PMF set remove_last = TRUE}
\usage{
evaluate_marg_cdf_simul_ci(marg_cdf_est, marg_cdf_eif, alpha, n,
  remove_last = FALSE)
}
\description{
this function could be renamed since it's used for CDF and PMF
for CDF, set remove_last = FALSE, for PMF set remove_last = TRUE
}
