% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_fn.R
\name{evaluate_marg_dist_simul_ci}
\alias{evaluate_marg_dist_simul_ci}
\title{Evaluate simultaneous confidence interval for marginal PMF or CDF.}
\usage{
evaluate_marg_dist_simul_ci(
  marg_dist_est,
  marg_dist_eif,
  alpha,
  remove_last = FALSE
)
}
\arguments{
\item{marg_dist_est}{The point estimate of the marginal CDF/PMF distribution}

\item{marg_dist_eif}{The EIF estimates for the marginal CDF/PMF estimates}

\item{alpha}{Confidence intervals have nominal level 1-\code{alpha}.}

\item{remove_last}{Should the last level be removed? Should be set equal to 
\code{TRUE} for CDF computations and \code{FALSE} for PMF computations.}
}
\value{
List by treatment of simultaneous confidence intervals
}
\description{
Evaluate simultaneous confidence interval for marginal PMF or CDF.
}
