% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_fn.R
\name{evaluate_marg_pmf_ptwise_ci}
\alias{evaluate_marg_pmf_ptwise_ci}
\title{Evaluate pointwise confidence interval for marginal PMF.}
\usage{
evaluate_marg_pmf_ptwise_ci(marg_pmf_est, marg_pmf_eif, alpha)
}
\arguments{
\item{marg_pmf_est}{The point estimate of the marginal PMF distribution}

\item{marg_pmf_eif}{The EIF estimates for the marginal PMF estimates}

\item{alpha}{Confidence intervals have nominal level 1-\code{alpha}.}
}
\value{
List by treatment of simultaneous confidence intervals
}
\description{
Evaluate pointwise confidence interval for marginal PMF.
}
