% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logodds_fn.R
\name{evaluate_theta_cov}
\alias{evaluate_theta_cov}
\title{get a covariance matrix for theta}
\usage{
evaluate_theta_cov(cdf_est, treat_prob_est, treat, out, out_levels)
}
\description{
get a covariance matrix for theta
}
