% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_fn.R
\name{fit_trt_spec_reg}
\alias{fit_trt_spec_reg}
\title{Helper function to fit a treatment specific outcome regression. 
If there are more than 2 observed levels of the outcome for the
specified treatment arm, then \code{polr} is used from the \code{MASS}
package. Otherwise logistic regression is used. In both cases, 
inverse probability of treatment weights are included in the regression. 
If there are levels of the outcome that are not observed in this treatment 
group, then 0's are added in. The function returns a matrix with named columns
corresponding to each outcome (ordered numerically). The entries 
represent the estimated covariate-conditional treatment-specific PMF.}
\usage{
fit_trt_spec_reg(trt_level, trt_spec_prob_est, out, treat, covar,
  out_levels, out_form = NULL, out_model, stratify, ...)
}
\arguments{
\item{trt_spec_prob_est}{A vector of estimates of Pr(\code{treat} = \code{trt_level} | \code{covar}).}

\item{out}{}
}
\description{
Helper function to fit a treatment specific outcome regression. 
If there are more than 2 observed levels of the outcome for the
specified treatment arm, then \code{polr} is used from the \code{MASS}
package. Otherwise logistic regression is used. In both cases, 
inverse probability of treatment weights are included in the regression. 
If there are levels of the outcome that are not observed in this treatment 
group, then 0's are added in. The function returns a matrix with named columns
corresponding to each outcome (ordered numerically). The entries 
represent the estimated covariate-conditional treatment-specific PMF.
}
