% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_fn.R
\name{get_one_marg_cdf}
\alias{get_one_marg_cdf}
\title{rename to reflect that its used for both CDF and PMF now}
\usage{
get_one_marg_cdf(treat, covar, treat_form, out, out_levels, out_form)
}
\description{
rename to reflect that its used for both CDF and PMF now
}
