% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooled_logistic.R
\name{getResponseFromFormula}
\alias{getResponseFromFormula}
\title{Get a response from model formula}
\usage{
getResponseFromFormula(formula, data)
}
\arguments{
\item{formula}{The model formula}
}
\description{
Get a response from model formula
}
