% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_fn.R
\name{marginalize_cdf}
\alias{marginalize_cdf}
\title{Marginalize over empirical distribution to obtain marginal
treatment-specific CDF estimate.}
\usage{
marginalize_cdf(cdf_est)
}
\arguments{
\item{cdf_est}{Estimates of treatment-specific conditional CDF.}
}
\description{
Marginalize over empirical distribution to obtain marginal
treatment-specific CDF estimate.
}
