% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_fn.R
\name{marginalize_pmf}
\alias{marginalize_pmf}
\title{Marginalize over empirical distribution to obtain marginal
treatment-specific PMF estimate.}
\usage{
marginalize_pmf(pmf_est)
}
\arguments{
\item{pmf_est}{Estimates of treatment-specific conditional PMF.}
}
\description{
Marginalize over empirical distribution to obtain marginal
treatment-specific PMF estimate.
}
