% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooled_logistic.R
\name{predict.POplugin}
\alias{predict.POplugin}
\title{Predict method for a \code{POplugin} object}
\usage{
\method{predict}{POplugin}(object, newdata = NULL)
}
\arguments{
\item{object}{An object of class \code{POplugin}}

\item{newdata}{A \code{data.frame} on which to predict}
}
\value{
A data frame with nrow = number of rows in newdata (or the orignal data frame)
and with the number of columns equal to the number of levels of the outcome observed in the original data frame
}
\description{
Predict method for a \code{POplugin} object
}
