% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.drord}
\alias{print.drord}
\title{Print the output of a \code{"drord"} object.}
\usage{
\method{print}{drord}(x, ci = "bca", ...)
}
\arguments{
\item{x}{A \code{"drord"} object}

\item{ci}{Which confidence interval should be printed. Defaults to BCa,
but it BCa was not computed in call to \code{drord}, defaults back to Wald.}

\item{...}{Other arguments (not used)}
}
\description{
Print the output of a \code{"drord"} object.
}
