% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmean_fn.R
\name{wald_ci_wmean}
\alias{wald_ci_wmean}
\title{Compute a Wald confidence interval for the weighted mean}
\usage{
wald_ci_wmean(wmean_est, alpha)
}
\arguments{
\item{wmean_est}{The estimated weighted means + estimated covariance matrix.}

\item{alpha}{Level of confidence interval.}
}
\value{
matrix with treatment-specific weighted mean CIs and CI for difference.
}
\description{
Compute a Wald confidence interval for the weighted mean
}
