% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_estimation.R
\name{drord}
\alias{drord}
\title{Doubly robust estimates of parameters for evaluating the effects 
of treatments on ordinal outcomes.}
\usage{
drord(out, treat, covar, out_levels = sort(unique(out)),
  out_form = ".", out_weights = rep(1, length(out_levels)),
  out_model = c("polr", "vglm", "clm"), treat_form = "1",
  param = c("weighted_mean", "log_odds", "mann_whitney"), ci = c("bca",
  "wald"), alpha = 0.05, test = c("wald"), nboot = 10000,
  return_models = TRUE, est_dist = TRUE, stratify = FALSE, ...)
}
\arguments{
\item{out}{A \code{numeric} vector containing the outcomes.}

\item{treat}{A \code{numeric} vector containing treatment status. Should only assume 
a value 0 or 1.}

\item{covar}{A \code{data.frame} containing the covariates to include in the working
proportional odds model.}

\item{out_levels}{A \code{numeric} vector containing all ordered levels of the 
outcome.}

\item{out_form}{The right-hand side of a regression formula for the working proportional 
odds model. NOTE: THIS FORMULA MUST NOT SUPPRESS THE INTERCEPT.}

\item{out_weights}{A vector of \code{numeric} weights with length equal to the length 
of \code{out_levels}.}

\item{out_model}{Which R function should be used to fit the proportional odds 
model. Options are \code{"polr"} (from the \code{MASS} package), 
"vglm" (from the \code{VGAM} package), or \code{"clm"} (from the \code{ordinal} package).}

\item{treat_form}{The right-hand side of a regression formula for the working model of
treatment probability as a function of covariates}

\item{param}{A vector of \code{characters} indicating which of the three treatment
effect parameters should be estimated (\code{"weighted_mean", "log_odds",} 
and/or \code{"mann_whitney"}).}

\item{ci}{A vector of \code{characters} indicating which confidence intervals
should be computed (\code{"bca"} and/or \code{"wald"})}

\item{alpha}{Confidence intervals have nominal level 1-\code{alpha}.}

\item{test}{A vector of \code{characters} indicating which test should be performed
(currently limited to \code{"wald"}).}

\item{nboot}{Number of bootstrap replicates used to compute bootstrap confidence
intervals.}

\item{return_models}{If \code{TRUE} the fitted working proportional odds models
and treatment probability models are returned.}

\item{est_dist}{A \code{boolean} indicating whether estimates of the CDF and PMF 
should be computed and returned. For real data analysis, we generally recommend 
leaving as \code{TRUE}; however, when studying performance in simulations, it can 
save time to set to \code{FALSE}.}

\item{stratify}{If \code{TRUE}, then a fully stratified estimator is computed, i.e.,
the empirical CDF of each treatment arm is estimated stratifying by levels of 
\code{covar}. For now, this option is limited to univariate covariates.}
}
\value{
TO DO: format output and add print/plot methods
}
\description{
The available parameters for evaluating treatment efficacy are:
\itemize{
\item Difference in (weighted) means: The outcome levels are treated
numerically, with each level possibly assigned a weight. The difference
in average outcomes is computed.
\item Log odds ratio: The comparison describes the average log-odds (treatment
level 1 versus 0) of the cumulative probability for each level of the outcome.
\item Mann-Whitney: The probability that a randomly-selected individual 
receiving treatment 1 will have a larger outcome value than a randomly selected
individual receiving treatment 0 (with ties assigned weight 1/2). 
}
}
\details{
In each case, estimates are constructed by obtaining a doubly robust 
estimate of the cumulative distribution function (CDF) for each treatment group. 
This is achieved by fitting a (working) proportional odds model that includes
inverse probability of treatment weights. The inclusion of these weights 
ensures that, so long as the working model includes intercept terms,
the resultant estimate of the CDF is an augmented inverse 
probability of treatment weighted estimate. This implies that the estimate is
nonparametric efficient if the working model contains the truth; however, 
even if the working model does not contain the truth, the CDF estimates are 
consistent and asymptotically normal with variance expected to dominate that
of an unadjusted estimate of the same treatment effect.

The CDF estimates are subsequently mapped into estimates of each requested 
parameter for evaluating treatment effects. The double robustness and efficiency 
properties of the CDF estimates extend to these quantities as well. Confidence
intervals and hypothesis tests can be carried out in closed form using Wald-style
intervals and tests or using a nonparametric corrected and accelerated bootstrap 
(BCa). Inference for the CDF and probability mass function is also returned and 
can be used for subsequent visualizations (see \code{plot.drord}).
}
\examples{
n <- 500
covar <- data.frame(x1 = runif(n), x2 = runif(n))
treat <- rbinom(n, 1, 1/2)
out <- rbinom(n, 3, plogis(-1 + covar$x1 - covar$x2 + 0.2 * treat))
# call drord and obtain wald inference
fit <- drord(out, treat, covar, ci = "wald")

# stratified estimator
covar <- data.frame(x1 = rbinom(n, 2, 0.25))
out <- rbinon(n, 3, plogis(-1 + covar$x1 + 0.5 * treat))
stratified_fit <- drord(out, treat, covar, ci = "wald", stratify = TRUE)
}
