% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logodds_fn.R
\name{evaluate_trt_spec_theta_eif}
\alias{evaluate_trt_spec_theta_eif}
\title{get a matrix of eif estimates for theta}
\usage{
evaluate_trt_spec_theta_eif(trt_spec_cdf_est, trt_spec_prob_est, trt_level,
  treat, out, out_levels)
}
\description{
get a matrix of eif estimates for theta
}
